move generate_password function
