+++ src/eduid/common/utils.py
@@ -3,6 +3,7 @@ __author__ = "lundberg"
 from uuid import uuid4
 
 from bson import ObjectId
+from pwgen import pwgen
 
 
 def urlappend(base: str, path: str) -> str:
@@ -52,3 +53,7 @@ def make_etag(version: ObjectId):
 
 def get_short_hash(entropy: int = 10) -> str:
     return uuid4().hex[:entropy]
+
+
+def generate_password(length: int = 12) -> str:
+    return pwgen(int(length), no_capitalize=True, no_symbols=True)
