+++ src/eduid/webapp/security/helpers.py
@@ -13,7 +13,7 @@ from eduid.userdb.security import SecurityUser
 from eduid.userdb.user import User
 from eduid.webapp.common.api.helpers import send_mail, set_user_names_from_official_address
 from eduid.webapp.common.api.messages import FluxData, TranslatableMsg, error_response
-from eduid.webapp.common.authn.utils import generate_password
+from eduid.common.utils import generate_password
 from eduid.webapp.common.session.namespaces import SP_AuthnRequest
 from eduid.webapp.security.app import current_security_app as current_app
